requirements.txt
# Core dependencies
click>=8.0.0
requests>=2.28.0
PyYAML>=6.0
python-dateutil>=2.8.0

# Async support
aiohttp>=3.8.0
asyncio

# Logging and monitoring
prometheus-client>=0.14.0

# Security
cryptography>=3.4.8

# Data validation
pydantic>=1.10.0

# Development and testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
requests-mock>=1.10.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Optional dependencies for enhanced features
python-dotenv>=0.19.0  # For loading .env files
Faker>=15.0.0  # For test data generation
tabulate>=0.9.0  # For formatted table output